backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic-settings
httpx
qrcode[pil]
Pillow
python-multipart
pytest
pytest-asyncio
niimprint
